# openapi.yml
...
openapi: 3.0.3
info:
  title: Uplink Connect API
  description: |
    Cloudflare-native data ingestion platform API.
    
    ## Authentication
    - **External endpoints**: Bearer token in `Authorization` header
    - **Internal endpoints**: `x-uplink-internal-key` header
  version: 3.01.0
  contact:
    name: BoringWorks
    url: https://github.com/Boring-Works/uplink-connect
servers:
  - url: https://uplink-core.codyboring.workers.dev
    description: Production Core API
  - url: https://uplink-edge.codyboring.workers.dev
    description: Production Edge API
...
paths:
...
  /internal/artifacts/{artifactId}:
    get:
      tags: [Artifacts]
      summary: Get artifact
      parameters:
        - name: artifactId
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Artifact details
        404:
          description: Artifact not found

  # Alerts
  /internal/alerts:
    get:
      tags: [Alerts]
      summary: List active alerts
      parameters:
        - name: severity
          in: query
          schema: { type: string, enum: [critical, warning, info] }
        - name: type
          in: query
          schema: { type: string }
        - name: sourceId
          in: query
          schema: { type: string }
        - name: acknowledged
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        200:
          description: List of alerts
  /internal/alerts/check:
    post:
      tags: [Alerts]
      summary: Run alert checks
      parameters:
        - name: sourceId
          in: query
          schema: { type: string }
      responses:
        200:
          description: Check results
  /internal/alerts/{alertId}/acknowledge:
    post:
      tags: [Alerts]
      summary: Acknowledge alert
      parameters:
        - name: alertId
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Alert acknowledged
        404:
          description: Alert not found
  /internal/alerts/{alertId}/resolve:
    post:
      tags: [Alerts]
      summary: Resolve alert
      parameters:
        - name: alertId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
      responses:
        200:
          description: Alert resolved
        404:
          description: Alert not found

  # Metrics
...
  /internal/browser/status:
    get:
      tags: [Browser]
      summary: Get browser manager status
      responses:
        200:
          description: Browser manager status
        502:
          description: Failed to get status
  /internal/notifications/test/{channel}:
    post:
      tags: [Browser]
      summary: Test notification channel
      parameters:
        - name: channel
          in: path
          required: true
          schema: { type: string, enum: [webhook, slack] }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string }
      responses:
        200:
          description: Test succeeded
        400:
          description: Test failed

  # Edge API (uplink-edge)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    InternalKey:
      type: apiKey
      in: header
      name: x-uplink-internal-key
security:
  - InternalKey: []